The problem causing the following problem:
"SQL State: 22008
Native Error: 0
Message Text: [Microsoft][ODBC Driver 18 for SQL Server]Datetime field overflow. Fractional second precision exceeds the scale specified in the parameter binding."

is fixed.

The fractions must have '00' (two zeros) at the end in order to set the nanoseconds properly with the precision set in the database. E.g. 100 represent the value of 0000001 and 999999900 represent the value of 9999999.